refactor: update Actions page structure
